move uploadRecordAction outside actions.ts to prevent circular import
