+++ src/lib/ai/question-generator.ts
@@ -1,9 +1,37 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import { Poppler } from "node-poppler";
-import { uploadRecordAction } from "@/src/app/actions";
 import type { StudyQuestion } from "@/src/types";
+import { createClient } from "../supabase/server";
 
+const DISPLAY_ORDER_STEP = 100;
+export async function uploadRecordAction(
+	upload_id: string,
+	questions: StudyQuestion[],
+	pos: number,
+) {
+	const supabase = await createClient();
+	// Insert questions
+	const questionRows = questions.map((q, idx) => ({
+		upload_id: upload_id,
+		question_text: q.question,
+		original_question_text: q.originalQuestion ?? q.question,
+		answer_text: q.answer,
+		original_answer_text: q.originalAnswer ?? q.answer,
+		page_number: q.pageNumber ?? null,
+		ocr_text: q.ocrText ?? null,
+		display_order: (pos + idx + 1) * DISPLAY_ORDER_STEP,
+		options: q.options ?? [],
+	}));
+
+	const { error } = await supabase
+		.from("questions")
+		.insert(questionRows as never);
+
+	if (error) {
+		console.error("Error inserting questions:", error);
+	}
+}
 //TODO optimize the pdf cario thing
 export async function generateQuestions(file: File, uploadId: string) {
 	if (!process.env.GEMINI_API_KEY) {
@@ -186,7 +214,7 @@ Return JSON array only:
 							ocrText: null,
 							originalQuestion: q.question,
 						}));
-						await uploadRecordAction(uploadId, questions);
+						await uploadRecordAction(uploadId, questions, i);
 					} catch (err) {
 						console.error("Error processing slide", {
 							fileName,
